backend/pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = ["xlartas <ivanhvalevskey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"

django-factory_boy = "1.0.0"
celery = "5.4.0"
celery-singleton = "0.3.1"
flower = "2.0.1"
redis = "5.0.8"
channels-redis = "4.2.0"
channels = "4.1.0"
daphne = "4.1.2"
twisted = {version = "24.7.0rc2", extras = ["http2", "tls"]}
httpx = "1.0.0b0"

Django = "5.0.8"
adrf = "0.1.7"
django-redis = "5.4.0"
django-cors-headers = "4.4.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
django-polymorphic = "3.1.0"
django-minio-backend = "3.6.0"
django-extensions = "3.2.3"
django-celery-beat = "2.6.0"
pytest-asyncio = "0.23.8"
pytest-django = "4.8.0"

aiohttp = "3.10.1"
chardet = "5.2.0"
python-dotenv = "1.0.1"

Pillow = "10.4.0"
psycopg2-binary = "2.9.9"
requests = "2.32.3"
setuptools = "71.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
